requirements.txt
solidpython
matplotlib
